removed include folder -> move header files to src
