+++ src/networks/metrics.cpp
@@ -4,7 +4,7 @@
 
 #include <numeric>
 
-#include "./networks/metrics.h"
+#include "metrics.h"
 
 template<class R, class... I>
 Meter<R, I...>::Meter(std::optional<int> window_size) : window_size(window_size), results() {
