+++ src/view/camera.h
